chore: move voting schemas to bulletin_board
